Move engine and lib components into separate files (#438)
Refactors library and engine structure to be less coupled. Pulls classes
and interfaces out into individual files. Cleans up implementation.
